collect parsing functions in parse.lua
and factor map-related functions out into map.lua
