+++ internal/quickfort/meta.lua
@@ -5,9 +5,8 @@ if not dfhack_flags.module then
     qerror('this script cannot be called directly')
 end
 
-local quickfort_common = reqscript('internal/quickfort/common')
-local log = quickfort_common.log
 local quickfort_command = reqscript('internal/quickfort/command')
+local quickfort_parse = reqscript('internal/quickfort/parse')
 
 -- blueprints referenced by meta blueprints must have a label, even if it's just
 -- a default numeric label. this is so we can provide intelligent error messages
@@ -19,7 +18,7 @@ local quickfort_command = reqscript('internal/quickfort/command')
 -- the lables are in ".csv format" even if they were fully specified for source
 -- .xlsx files, but those .xlsx files have since been serialized to .csv.
 local function get_section_name(cell, text, cur_sheet_name)
-    local sheet_name, label = quickfort_command.parse_section_name(text)
+    local sheet_name, label = quickfort_parse.parse_section_name(text)
     if not label then
         local cell_msg = string.format(
             'malformed blueprint section name in cell %s', cell)
@@ -43,7 +42,7 @@ local function do_meta(zlevel, grid, ctx)
             {label='Blueprints applied', value=0, always=true}
 
     -- ensure we process blueprints in the declared order
-    local cells = quickfort_common.get_ordered_grid_cells(grid)
+    local cells = quickfort_parse.get_ordered_grid_cells(grid)
     local saved_zlevel = ctx.cursor.z
     for _, cell in ipairs(cells) do
         local section_name =
