+++ internal/quickfort/notes.lua
@@ -5,10 +5,10 @@ if not dfhack_flags.module then
     qerror('this script cannot be called directly')
 end
 
-local quickfort_common = reqscript('internal/quickfort/common')
+local quickfort_parse = reqscript('internal/quickfort/parse')
 
 function do_run(zlevel, grid, ctx)
-    local cells = quickfort_common.get_ordered_grid_cells(grid)
+    local cells = quickfort_parse.get_ordered_grid_cells(grid)
     local lines = {}
     local prev_y = nil
     for _,cell in ipairs(cells) do
