+++ quickfort.lua
@@ -139,6 +139,7 @@ local quickfort_dialog = reqscript('internal/quickfort/dialog')
 local quickfort_dig = reqscript('internal/quickfort/dig')
 local quickfort_keycodes = reqscript('internal/quickfort/keycodes')
 local quickfort_list = reqscript('internal/quickfort/list')
+local quickfort_map = reqscript('internal/quickfort/map')
 local quickfort_meta = reqscript('internal/quickfort/meta')
 local quickfort_notes = reqscript('internal/quickfort/notes')
 local quickfort_orders = reqscript('internal/quickfort/orders')
